=== ChangeSet ===
Added functionality for interacting with DatePicker.
Getting the currently selected date in the Picker returning as string or Calendar.
Getting the individual componenets of the Picker as int instances.
Setting a Date on the current DatePicker by given Calendar.

=== InitialImports ===

Added method for getting the DatePicker in Screen
Added enum in ContentDescriptor for getting the DatePicker button.
Added class for manging DatePicker functionality like getting and setting value.
Added Integration test for DatePicker functionality.

[https://inttools.musala.com/reviewboard/r/31/ $31] see #3451

git-svn-id: https://trac.musala.com/ARDC/repo@10300 34708651-a5da-c34d-b20c-10ddab38317d